+++ packages/core/interface/src/common/interfaces/headers.ts
@@ -2,4 +2,5 @@ export const Headers = {
   X_VISITOR_ID: 'X-Visitor-Id',
   X_CAPTCHA_TOKEN: 'X-Captcha-Token',
   X_CAPTCHA_ID: 'X-Captcha-Identity',
+  X_API_VERSION: 'X-API-VERSION',
 };